Week8/day5/deploy/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
llama-cpp-python==0.2.77
requests==2.31.0
streamlit==1.35.0
python-multipart==0.0.9